requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
Flask==3.1.1
fonttools==4.59.0
fsspec==2024.6.1
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.8
lightning-utilities==0.14.3
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
scipy==1.15.3
six==1.17.0
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.4
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.9.0
tzdata==2025.2
Werkzeug==3.1.3
